Added a PWA manifest
Progressive web apps on iOS require an icon to be listed, but it doesn't seem
to require that the icon actually exists. For now it seems easiest to just use
a path to a non-existent image unless its found to be needed on other
platforms, or we have an icon we can use.